requirements.txt
youtube_transcript_api==0.6.1
streamlit==1.29.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pathlib==1.0.1